split signal module from stdlib
